+++ CHANGELOG.md
@@ -19,6 +19,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - CiscoCodecDirectSharingControl to support Cisco Proximity sharing
  - Added mask to ConferencePoint for configuring privacy mute authority
  - ConferenceManager uses privacy mute masks to better handle privacy mute state changes
+ 
+### Changed
+ - Cisco Codec Occupancy Control now reports people counts when available
 
 ## [18.0.0] - 2021-05-14
 ### Added